+++ src/main/java/com/qa/business/service/TraineeService.java
@@ -3,6 +3,6 @@ package com.qa.business.service;
 public interface TraineeService {
 	String getAll();
 	String createTrainee(String trainee);
-	String deleteTrainee(Long id);
-	String updateTrainee(Long id, String trainee);
+	String deleteTrainee(Integer id);
+	String updateTrainee(Integer id, String trainee);
 }
